Initielt oppsett av msw
